+++ src/signals/operators/signalize.ts
@@ -1,11 +1,10 @@
 import { Observable } from "rxjs";
 
-import { ReadableSignalFnLike } from "@/signals/types";
+import { SourceSignal } from "../base";
+import { type ReadonlySignal } from "../types";
 
-import { ReadonlySignal } from "../base";
-
-export function signalize<T>(observable: Observable<T>): ReadableSignalFnLike<T> {
-    return ReadonlySignal.create((destination) => {
-        return observable.subscribe(destination);
-    });
+export function signalize<T>(observable: Observable<T>): ReadonlySignal<T>;
+export function signalize<T>(observable: Observable<T>, defaultValue: T): ReadonlySignal<T>;
+export function signalize<T>(observable: Observable<T>, ...defaultValue: [defaultValue?: T]): ReadonlySignal<T> {
+    return SourceSignal.create((destination) => observable.subscribe(destination), ...defaultValue);
 }